# backend/services/rating-service/openapi.yaml
openapi: "3.0.3"
info:
  title: Clay Rating Service
  version: "1.0.0"
  description: |
    Rating and review management for Clay superapp.

    ## Responsibilities
    - Accepts post-trip ratings from users for drivers and merchants
    - Accepts post-trip ratings from drivers for passengers
    - Stores reviews (text + score)
    - Computes and caches average ratings per driver/merchant/passenger
    - Exposes average rating to other services (Matching, Merchant)

    ## Rating Subjects
    - **Driver**: rated by user after ride/delivery/food order
    - **Merchant**: rated by user after food order (alongside driver)
    - **Passenger**: rated by driver after ride order

    ## Score Scale
    1–5 stars (integer). Half-stars not supported.

    ## Storage
    - **PostgreSQL**: `ratings` table (subject_type, subject_id, order_id, rater_id, score, review_text)
    - **Redis**: `rating:avg:{subject_type}:{subject_id}` (cached average, TTL 1h)

    ## Constraints
    - One rating per (order_id, rater_id, subject_id) — enforced by unique index
    - Rating window: 24h after order completion (enforced by checking order timestamps)
servers:
  - url: http://clay-rating-service:8080
    description: Rating service
...
paths:
...
  /ratings:
    post:
      summary: Submit a rating
      operationId: submitRating
      tags: [Rating]
      description: |
        Called by order services (via proxy) after trip completion.
        Supports submitting multiple subjects in one call
        (e.g., driver + merchant for food orders).
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitRatingRequest"
      responses:
        "201":
          description: Rating submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingSubmitResponse"
        "409":
          description: Rating already submitted for this order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Rating window expired (> 24h after completion)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    UserId:
      name: X-User-ID
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Authenticated user ID (set by API Gateway from JWT)
  schemas:
...
    SubjectType:
      type: string
      enum: [driver, merchant, passenger]

    # ── Submit Rating ─────────────────────────────────────────────────────────
    SubmitRatingRequest:
      type: object
      required: [order_id, ratings]
      properties:
        order_id:
          type: string
          format: uuid
        ratings:
          type: array
          items:
            $ref: "#/components/schemas/RatingEntry"
          minItems: 1
          maxItems: 2
          description: One entry for simple orders; two entries (driver + merchant) for food orders
    RatingEntry:
      type: object
      required: [subject_type, subject_id, score]
      properties:
        subject_type:
          $ref: "#/components/schemas/SubjectType"
        subject_id:
          type: string
          format: uuid
        score:
          type: integer
          minimum: 1
          maximum: 5
        review_text:
          type: string
          maxLength: 500
          nullable: true
        tags:
          type: array
          items:
            type: string
            example: "friendly"
          description: Predefined quick-tag options (optional)
    RatingSubmitResponse:
      type: object
      properties:
        submitted:
          type: array
          items:
            type: object
            properties:
              rating_id:
                type: string
                format: uuid
              subject_type:
                $ref: "#/components/schemas/SubjectType"
              subject_id:
                type: string
                format: uuid
              score:
                type: integer

    # ── Get Ratings ───────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: ALREADY_RATED
        message:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT